Улучшил обработку ошибок, избавился от бага в логах, изменил сваггер

+++ internal/service/song.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"fmt"
 	"github.com/Xapsiel/EffectiveMobile/internal/models"
 	"github.com/Xapsiel/EffectiveMobile/internal/repository"
 	"strings"
@@ -18,6 +19,9 @@ func (s *SongService) GetSongs(filter models.Filter) ([]models.Song, error) {
 	return s.repo.GetSongs(filter)
 }
 func (s *SongService) GetSongVerse(song models.Song, VerseNumber int) (string, int, error) {
+	if song.ID == 0 && (song.SongName == "" || song.Group == "") {
+		return "", 0, fmt.Errorf("Не все данные были переданы")
+	}
 	text, id, err := s.repo.GetSongVerse(song)
 	if err != nil {
 		return "", 0, err
@@ -30,12 +34,21 @@ func (s *SongService) GetSongVerse(song models.Song, VerseNumber int) (string, i
 	return textParts[VerseNumber], id, nil
 }
 func (s *SongService) DeleteSong(song models.Song) (bool, error) {
+	if song.ID == 0 && (song.SongName == "" || song.Group == "") {
+		return false, fmt.Errorf("Не все данные были переданы")
+	}
 	return s.repo.DeleteSong(song)
 }
 func (s *SongService) UpdateSong(song models.Song) (bool, models.Song, error) {
+	if song.ID == 0 && (song.SongName == "" || song.Group == "") {
+		return false, models.Song{}, fmt.Errorf("Не все данные были переданы")
+	}
 	return s.repo.UpdateSong(song)
 }
 func (s *SongService) Add(song models.Song) (bool, int, error) {
+	if song.SongName == "" || song.Group == "" {
+		return false, 0, fmt.Errorf("Не все данные были переданы")
+	}
 	return s.repo.Add(song)
 }
